requirements.txt
absl-py==1.2.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
enum34==1.1.10
executing==0.10.0
fastjsonschema==2.16.1
flatbuffers==1.12
fonttools==4.36.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.12.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.1
keras==2.9.0
Keras-Applications==1.0.8
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plaidbench==0.7.0
plaidml==0.7.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.13.0
Pympler==1.0.1
pyopencl==2022.1.6
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytools==2022.1.12
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.4.0
streamlit==1.12.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.1
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.0.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.1
wrapt==1.12.1
zipp==3.8.1
